// src/utils/types/pathTypes.ts
export type FeaturesPath =
  | "integrations"
  | "paymentOrchestrationTools"
  | "merchantManagementTools"
  | "financialStatementAndBalanceManagement"
  | "antiFraudTools"
  | "offerYourMerchants"
  | "pricingOptions";

export type CompanyPath = "eComChangeAndTeam" | "PCIDSScertificate";

export type ResourcesPath = "blog" | "topQuestions" | "useCases" | "privacyPolicy" | "sitemap";

export type ContactsPath = "addressAndContact" | "contactSalesForm"| "getDemoAccessToTheBackOfficeForm"